[generator] Create base class for unit tests.
