--- tools/generator/Tests/Unit-Tests/JavaInteropCodeGeneratorTests.cs
@@ -1,11 +1,0 @@
-using NUnit.Framework;
-using Xamarin.Android.Binder;
-
-namespace generatortests
-{
-	[TestFixture]
-	class JavaInteropCodeGeneratorTests : CodeGeneratorTests
-	{
-		protected override CodeGenerationTarget Target => CodeGenerationTarget.JavaInterop1;
-	}
-}
