--- tools/generator/Tests/Unit-Tests/XamarinAndroidCodeGeneratorTests.cs
@@ -1,11 +1,0 @@
-using NUnit.Framework;
-using Xamarin.Android.Binder;
-
-namespace generatortests
-{
-	[TestFixture]
-	class XamarinAndroidCodeGeneratorTests : CodeGeneratorTests
-	{
-		protected override CodeGenerationTarget Target => CodeGenerationTarget.XamarinAndroid;
-	}
-}
