+++ tools/generator/Tests/generator-Tests.csproj
@@ -69,13 +69,12 @@
     <Compile Include="Integration-Tests\GenericArguments.cs" />
     <Compile Include="Integration-Tests\InterfaceMethodsConflict.cs" />
     <Compile Include="SupportFiles\RegisterAttribute.cs" />
+    <Compile Include="Unit-Tests\CodeGeneratorTestBase.cs" />
     <Compile Include="Unit-Tests\CodeGeneratorTests.cs" />
     <Compile Include="Unit-Tests\AdjusterTests.cs" />
-    <Compile Include="Unit-Tests\JavaInteropCodeGeneratorTests.cs" />
     <Compile Include="Unit-Tests\ManagedTests.cs" />
     <Compile Include="Unit-Tests\SupportTypes.cs" />
     <Compile Include="Unit-Tests\TestExtensions.cs" />
-    <Compile Include="Unit-Tests\XamarinAndroidCodeGeneratorTests.cs" />
     <Compile Include="Unit-Tests\XmlTests.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildBinPath)\Microsoft.CSharp.targets" />
